requirements.txt
BeautifulSoup4
selenium
lxml
discord.py
